% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRegistrationTracking.R
\name{fullRegistrationTrackingBoxModuleServer}
\alias{fullRegistrationTrackingBoxModuleServer}
\title{fullRegistrationTrackingBoxModuleServer}
\usage{
fullRegistrationTrackingBoxModuleServer(
  id,
  input,
  output,
  session,
  registrationDataBundle,
  classesStart
)
}
\arguments{
\item{id}{the ID}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{registrationDataBundle}{the input data}

\item{classesStart}{dates that classes start by semester}
}
\description{
fullRegistrationTrackingBoxModuleServer
}
