% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRegistrationTracking.R
\name{fullRegistrationTrackingBoxModuleUI}
\alias{fullRegistrationTrackingBoxModuleUI}
\title{fullRegistrationTrackingBoxModuleUI}
\usage{
fullRegistrationTrackingBoxModuleUI(id)
}
\arguments{
\item{id}{the ID}
}
\description{
fullRegistrationTrackingBoxModuleUI
}
