% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRegistrationTrackingDB.R
\name{fullRegistrationTrackingExternalDBBoxModuleServer}
\alias{fullRegistrationTrackingExternalDBBoxModuleServer}
\title{fullRegistrationTrackingExternalDBBoxModuleServer}
\usage{
fullRegistrationTrackingExternalDBBoxModuleServer(
  id,
  input,
  output,
  session,
  dbConn,
  classesStart,
  registrationDataBundle
)
}
\arguments{
\item{id}{the ID}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{dbConn}{the connection object for the external database}

\item{classesStart}{dates that classes start by semester}

\item{registrationDataBundle}{used to access full registration tracking until can be fixed}
}
\description{
fullRegistrationTrackingExternalDBBoxModuleServer
}
