% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRegistrationTrackingDB.R
\name{fullRegistrationTrackingExternalDBBoxModuleUI}
\alias{fullRegistrationTrackingExternalDBBoxModuleUI}
\title{fullRegistrationTrackingExternalDBBoxModuleUI}
\usage{
fullRegistrationTrackingExternalDBBoxModuleUI(id)
}
\arguments{
\item{id}{the ID}
}
\description{
fullRegistrationTrackingExternalDBBoxModuleUI
}
