% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicEnrollment.R
\name{historicEnrollmentServer}
\alias{historicEnrollmentServer}
\title{historicEnrollmentServer}
\usage{
historicEnrollmentServer(id, inData, inTrackingData, semester.codes)
}
\arguments{
\item{id}{the id}

\item{inData}{the input data}

\item{inTrackingData}{the tracking data}

\item{semester.codes}{the semester.codes}
}
\description{
historicEnrollmentServer
}
