% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicEnrollment.R
\name{historicEnrollmentUI}
\alias{historicEnrollmentUI}
\title{historicEnrollmentUI}
\usage{
historicEnrollmentUI(id, label = "Historic Enrollment")
}
\arguments{
\item{id}{the id}

\item{label}{the label}
}
\description{
historicEnrollmentUI
}
