% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingOrig.R
\name{registrationTrackingOrigServer}
\alias{registrationTrackingOrigServer}
\title{registrationTrackingOrigServer}
\usage{
registrationTrackingOrigServer(
  id,
  deptAbbrv,
  focalSem,
  refSem,
  ugControl,
  syncSwitch,
  sameTerm,
  chosenCourse,
  previousSemestersFinalEnrollment,
  fullRegistrationTracking,
  useShort = FALSE,
  semester.codes,
  classesStart,
  synonyms
)
}
\arguments{
\item{id}{the id}

\item{deptAbbrv}{the Department Abbreviation}

\item{focalSem}{focal Semester}

\item{refSem}{reference Semestr}

\item{ugControl}{undergraduate control}

\item{syncSwitch}{syncronize switch}

\item{sameTerm}{same term}

\item{chosenCourse}{chosen course}

\item{previousSemestersFinalEnrollment}{previous semester's final enrollment}

\item{fullRegistrationTracking}{full registration tracking data}

\item{useShort}{use short data format}

\item{semester.codes}{the semester codes}

\item{classesStart}{list of dates when classes start by semester}

\item{synonyms}{for departments (e.g. PSYC=PBSI)}
}
\description{
registrationTrackingOrigServer
}
