% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingOrig.R
\name{registrationTrackingOrigUI}
\alias{registrationTrackingOrigUI}
\title{registrationTrackingOrigUI}
\usage{
registrationTrackingOrigUI(id, label = "Registration Tracking")
}
\arguments{
\item{id}{the ID}

\item{label}{the Label}
}
\description{
registrationTrackingOrigUI
}
