% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingSummary.R
\name{registrationTrackingSummaryServer}
\alias{registrationTrackingSummaryServer}
\title{registrationTrackingSummaryServer}
\usage{
registrationTrackingSummaryServer(id, inTrackingData, ppData, semester.codes)
}
\arguments{
\item{id}{the id}

\item{inTrackingData}{the tracking data}

\item{ppData}{the ppData}

\item{semester.codes}{the semester codes}
}
\description{
registrationTrackingSummaryServer
}
