% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingSummary.R
\name{registrationTrackingSummaryUI}
\alias{registrationTrackingSummaryUI}
\title{registrationTrackingSummaryUI}
\usage{
registrationTrackingSummaryUI(id, label = "Registration Tracking")
}
\arguments{
\item{id}{the id}

\item{label}{the label}
}
\description{
registrationTrackingSummaryUI
}
