% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicEnrollment.R
\name{historicEnrollmentServer}
\alias{historicEnrollmentServer}
\title{historicEnrollmentServer}
\usage{
historicEnrollmentServer(
  id,
  inData,
  inTrackingData,
  deptAbbrv,
  ugControl,
  chosenCourse,
  semester.codes,
  synonyms
)
}
\arguments{
\item{id}{the id}

\item{inData}{the input data}

\item{inTrackingData}{the tracking data}

\item{deptAbbrv}{the department abbreviation}

\item{ugControl}{the undergraduate control}

\item{chosenCourse}{the chosen course}

\item{semester.codes}{the semester.codes}

\item{synonyms}{table of synonyms for department codes (e.g. PSYC = PBSI)}
}
\description{
historicEnrollmentServer
}
