% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingOrig.R
\name{registrationTrackingOrigServer}
\alias{registrationTrackingOrigServer}
\title{registrationTrackingOrigServer}
\usage{
registrationTrackingOrigServer(
  id,
  deptAbbrv,
  previousSemestersFinalEnrollment,
  fullRegistrationTracking,
  useShort = FALSE,
  semester.codes
)
}
\arguments{
\item{id}{the id}

\item{deptAbbrv}{the Department Abbreviation}

\item{previousSemestersFinalEnrollment}{previous semester's final enrollment}

\item{fullRegistrationTracking}{full registration tracking data}

\item{useShort}{use short data format}

\item{semester.codes}{the semester codes}
}
\description{
registrationTrackingOrigServer
}
