% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationTrackingSummary.R
\name{registrationTrackingSummaryServer}
\alias{registrationTrackingSummaryServer}
\title{registrationTrackingSummaryServer}
\usage{
registrationTrackingSummaryServer(
  id,
  inTrackingData,
  ppData,
  semester.codes,
  focalSem,
  refSem,
  deptAbbrv,
  classesStart,
  synonyms
)
}
\arguments{
\item{id}{the id}

\item{inTrackingData}{the tracking data}

\item{ppData}{the ppData}

\item{semester.codes}{the semester codes}

\item{focalSem}{focalSem}

\item{refSem}{refSem}

\item{deptAbbrv}{deptAbbrv}

\item{classesStart}{classesStart}

\item{synonyms}{synonyms}
}
\description{
registrationTrackingSummaryServer
}
